app4/facade-service/requirements.txt
fastapi
uvicorn
pydantic
httpx
hazelcast-python-client==5.4.0
python-consul